add test suite utility
