+++ test/utils.ts
@@ -3,6 +3,8 @@ import * as path from "node:path";
 import * as url from "node:url";
 
 import * as glob from "glob";
+import { z } from "zod";
+import camelcase from "camelcase";
 
 let __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 let fixtures = path.join(__dirname, "..", "ngff");
@@ -15,9 +17,44 @@ export function globber(version: string) {
     return files.map((file) => ({
       name: path.basename(file),
       path: file,
-      json() {
-        return fs.readFile(file, { encoding: "utf-8" }).then(JSON.parse);
-      },
+      json: () => fs.readFile(file, { encoding: "utf-8" }).then(JSON.parse),
     }));
   };
+}
+
+let Test = z.object({
+  formerly: z.string(),
+  data: z.any(),
+  valid: z.boolean(),
+});
+
+let Suite = z.object({
+  description: z.string(),
+  schema: z.object({
+    id: z.string(),
+  }),
+  tests: z.array(Test),
+});
+
+export async function gather_test_cases<T extends Record<string, z.ZodSchema>>(
+  version: "0.4" | "latest",
+  schemas: T,
+) {
+  let files = await glob.glob(
+    path.join(__dirname, "..", "ngff", version, "/tests/*.json"),
+  );
+
+  let cases = files.map(async (file) => {
+    let suite = await fs.readFile(file, "utf-8")
+      .then(JSON.parse)
+      .then(Suite.parse);
+
+    let cased = camelcase(suite.schema.id.split("/").pop()!);
+    let name = cased.charAt(0).toUpperCase() +
+      cased.slice(1) as keyof typeof schemas;
+
+    return Object.assign(suite, { Schema: schemas[name] });
+  });
+
+  return Promise.all(cases);
 }
